# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Water System SCADA + WebGIS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ValveStatus = (typeof ValveStatus)[keyof typeof ValveStatus];

export const ValveStatus = {
  normal: "normal",
  warning: "warning",
  critical: "critical",
} as const;

export interface Valve {
  id: number;
  valveId: string;
  name: string;
  lat: number;
  lng: number;
  pressure: number;
  status: ValveStatus;
  createdAt?: string;
  updatedAt?: string;
}

export interface CreateValveBody {
  valveId: string;
  name: string;
  lat: number;
  lng: number;
  pressure: number;
}

export interface UpdateValveBody {
  name?: string;
  lat?: number;
  lng?: number;
  pressure?: number;
}

export interface Pipe {
  id: number;
  name: string;
  diameter?: number | null;
  material?: string | null;
  fromNode?: string | null;
  toNode?: string | null;
  coordinates: number[][];
  createdAt?: string;
  updatedAt?: string;
}

export interface CreatePipeBody {
  name: string;
  diameter?: number;
  material?: string;
  fromNode?: string;
  toNode?: string;
  coordinates: number[][];
}

export interface UpdatePipeBody {
  name?: string;
  diameter?: number;
  material?: string;
  fromNode?: string;
  toNode?: string;
  coordinates?: number[][];
}

export interface WaterSource {
  id: number;
  name: string;
  lat: number;
  lng: number;
  createdAt?: string;
}

export interface CreateSourceBody {
  name: string;
  lat: number;
  lng: number;
}

export type GeoJsonImportFeaturesItem = { [key: string]: unknown };

export interface GeoJsonImport {
  type: string;
  features: GeoJsonImportFeaturesItem[];
}

export interface ImportResult {
  success: boolean;
  valvesImported: number;
  pipesImported: number;
  message: string;
}

export interface TelemetryData {
  valveId: string;
  pressure: number;
}

export interface TelemetryResponse {
  success: boolean;
  valve?: Valve;
}

export interface DashboardStats {
  totalValves: number;
  totalPipes: number;
  totalSources: number;
  avgPressure: number;
  normalCount: number;
  warningCount: number;
  criticalCount: number;
}

export interface PressureRecord {
  valveId: string;
  name: string;
  pressure: number;
  timestamp: string;
}

export type ExportGeoJson200 = { [key: string]: unknown };