# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Water System SCADA + WebGIS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /valves:
    get:
      operationId: listValves
      tags: [valves]
      summary: List all valves
      responses:
        "200":
          description: List of valves
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Valve"
    post:
      operationId: createValve
      tags: [valves]
      summary: Create a new valve
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateValveBody"
      responses:
        "201":
          description: Valve created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Valve"
  /valves/{id}:
    get:
      operationId: getValve
      tags: [valves]
      summary: Get valve by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Valve details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Valve"
    put:
      operationId: updateValve
      tags: [valves]
      summary: Update a valve
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateValveBody"
      responses:
        "200":
          description: Valve updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Valve"
    delete:
      operationId: deleteValve
      tags: [valves]
      summary: Delete a valve
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Valve deleted
...
components:
  schemas:
...
    Valve:
      type: object
      properties:
        id:
          type: integer
        valveId:
          type: string
        name:
          type: string
        lat:
          type: number
        lng:
          type: number
        pressure:
          type: number
        status:
          type: string
          enum: [normal, warning, critical]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - valveId
        - name
        - lat
        - lng
        - pressure
        - status
    CreateValveBody:
      type: object
      properties:
        valveId:
          type: string
        name:
          type: string
        lat:
          type: number
        lng:
          type: number
        pressure:
          type: number
      required:
        - valveId
        - name
        - lat
        - lng
        - pressure
    UpdateValveBody:
      type: object
      properties:
        name:
          type: string
        lat:
          type: number
        lng:
          type: number
        pressure:
          type: number
...