# lib/api-zod/src/generated/types/updateValveBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Water System SCADA + WebGIS API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateValveBody {
  name?: string;
  lat?: number;
  lng?: number;
  pressure?: number;
}