# lib/api-zod/src/generated/types/valve.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Water System SCADA + WebGIS API
 * OpenAPI spec version: 0.1.0
 */
import type { ValveStatus } from "./valveStatus";

export interface Valve {
  id: number;
  valveId: string;
  name: string;
  lat: number;
  lng: number;
  pressure: number;
  status: ValveStatus;
  createdAt?: Date;
  updatedAt?: Date;
}